LayoutTests/platform/ios/accessibility/dynamically-changing-iframe-remains-accessible-expected.txt

This test ensures that an iframe that changes its document dynamically remains accessible.

Beginning search traversal from the body element.

TextField
text-input-outside-iframe

StaticText
AXLabel: Purple pineapple

StaticText
AXLabel: Purple pineapple

Traversed 3 elements.
PASS successfullyParsed is true

TEST COMPLETE
